+++ app/src/navigation/router.tsx
@@ -7,6 +7,7 @@ import SurveyNavigator from "../scenes/survey-v2/SurveyNavigator";
 import SelectDayScreen from "../scenes/survey/selectDay";
 import Reminder from "../scenes/reminder";
 import Export from "../scenes/export/export";
+import DataExportImport from "../scenes/data-export-import";
 import DailyChart from "../scenes/calendar/daily-chart";
 import { AppState, Platform, Linking } from "react-native";
 import { StatusBar } from "expo-status-bar";
@@ -72,6 +73,8 @@ import { StatusBarProvider, useStatusBarInternal } from "../context/StatusBarCon
 import { TW_COLORS } from "@/utils/constants";
 import SurveyV1 from "../scenes/survey/daySurvey";
 import SurveySuccessScreen from "../scenes/survey/SurveySuccessScreen";
+import FaqMainScreen from "@/scenes/faq/FaqMainScreen";
+import FaqDetailScreen from "@/scenes/faq/FaqDetailScreen";
 
 const Stack = createStackNavigator();
 
@@ -218,6 +221,8 @@ class Router extends React.Component<RouterProps> {
             }}
           >
             <Stack.Screen name="presentation" component={Presentation} />
+            <Stack.Screen name="faq" component={FaqMainScreen} />
+            <Stack.Screen name="faq-detail" component={FaqDetailScreen} />
             <Stack.Screen name="day-survey" component={SurveyV1} />
             <Stack.Screen name="day-survey-v2" component={SurveyNavigator} />
             <Stack.Screen name="survey-success" component={SurveySuccessScreen} />
@@ -226,6 +231,7 @@ class Router extends React.Component<RouterProps> {
             <Stack.Screen name="symptoms" component={Indicateurs} />
             <Stack.Screen name="reminder" component={Reminder} />
             <Stack.Screen name="export" component={Export} />
+            <Stack.Screen name="data-export-import" component={DataExportImport} />
             <Stack.Screen name="chart-day" component={DailyChart} />
             <Stack.Screen name="notes" options={{ animationEnabled: Platform.OS === "ios" }}>
               {({ navigation, route }) => <Notes navigation={navigation} route={route} />}