+++ CHANGELOG.md
@@ -1,3 +1,45 @@
+# [1.70.0](https://github.com/SocialGouv/jardinmental/compare/v1.69.0...v1.70.0) (2025-09-16)
+
+
+### Features
+
+* **DiaryList:** add mood selector for current day at top of DiaryList ([#446](https://github.com/SocialGouv/jardinmental/issues/446)) ([5f926ee](https://github.com/SocialGouv/jardinmental/commit/5f926ee4f9668c990e5ee7741be3e91ef1e175fa))
+
+# [1.69.0](https://github.com/SocialGouv/jardinmental/compare/v1.68.1...v1.69.0) (2025-09-16)
+
+
+### Features
+
+* add goal deactivation ([#457](https://github.com/SocialGouv/jardinmental/issues/457)) ([7d6c77b](https://github.com/SocialGouv/jardinmental/commit/7d6c77b631eb2431d9748bc6c4dce8e3abd78522))
+
+## [1.68.1](https://github.com/SocialGouv/jardinmental/compare/v1.68.0...v1.68.1) (2025-09-16)
+
+
+### Bug Fixes
+
+* **daysurvey:** fix padding flickering when daysurvey load ([#460](https://github.com/SocialGouv/jardinmental/issues/460)) ([bd671fc](https://github.com/SocialGouv/jardinmental/commit/bd671fcd35c8bae54de7637268e8aceca8713c16))
+
+# [1.68.0](https://github.com/SocialGouv/jardinmental/compare/v1.67.0...v1.68.0) (2025-09-16)
+
+
+### Features
+
+* **daysurvey:** on android change status bar according to selected mood ([#459](https://github.com/SocialGouv/jardinmental/issues/459)) ([fef9628](https://github.com/SocialGouv/jardinmental/commit/fef96281bc41b62133b2b465f00ecd9dc074f1ac))
+
+# [1.67.0](https://github.com/SocialGouv/jardinmental/compare/v1.66.0...v1.67.0) (2025-09-16)
+
+
+### Features
+
+* improve UI with icon updates and styling fixes ([#458](https://github.com/SocialGouv/jardinmental/issues/458)) ([ad7ab87](https://github.com/SocialGouv/jardinmental/commit/ad7ab870829bd3ea69a95826c590037b4a0e3675))
+
+# [1.66.0](https://github.com/SocialGouv/jardinmental/compare/v1.65.0...v1.66.0) (2025-09-15)
+
+
+### Features
+
+* add import/export ([#450](https://github.com/SocialGouv/jardinmental/issues/450)) ([ce4175d](https://github.com/SocialGouv/jardinmental/commit/ce4175d2a1a8c38d5a5e1ceae3e39ce65fdfcc98))
+
 # [1.65.0](https://github.com/SocialGouv/jardinmental/compare/v1.64.0...v1.65.0) (2025-09-15)